pyproject.toml
[project]
name = "livekit-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.4.1",
    "livekit-agents>=1.3.6",
    "livekit-plugins-google>=1.3.6",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.3.6",
    "livekit-plugins-silero>=1.3.6",
    "mem0ai>=1.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
